Reorder and add vulnerabilities check in PHP workflow
The PHP workflow of the GitHub Actions has been adjusted. Composer validation is now done after caching and before the installation of dependencies. A new step for checking vulnerabilities has been added. This step is run if the installation of dependencies is successful, before the parallel linting and test runs.