Remove usage of FinalizeClassesWithoutChildrenRector
The usage of FinalizeClassesWithoutChildrenRector class has been removed from the rector.php file. This was done because its functionality was causing bugs due to incomplete knowledge of the full dependency tree, making it risky to use. A suggested substitute is the "Finalize" class from TomasVotruba on GitHub, which operates with full context.