Add comparison link to release notes
A comparison link has been added to the release notes in GitHub Actions workflow. This link redirects to a page showing the differences between the last and the new tag. The comparison link is only created if the previous steps were successful and the `LATEST_TAG` environment variable is set.