Refactor commit message generation in GitHub workflow
The commit is a simplification of the commands to get the commit messages. The process of generating commit messages in the '.github/workflows/php.yml' has been streamlined by replacing the complex group of commands with a simpler 'git log' command. This change should make the workflow more efficient and easier to read.