Update composer.json dependencies versioning
Changed version constraints for development dependencies in composer.json file from specific versions to be compatible with versions equal to or greater than the current ones. This modification aims to utilize the latest available versions of those dependencies without violating the established constraints, ensuring the project stays up-to-date with the changes in these dependencies.