Update PHP workflow settings in Github
Updated the .github/workflows/php.yml file to enhance the workflow. Added a 'fail-fast: false'flag to the matrix strategy section to prevent other jobs from stopping if one job fails. Also, added a condition 'if: steps.composer-install.outcome =='success' under the vulnerabilities-check to ensure it only runs if the composer installation was successful.