(Fixed) Refactor release creation conditions in GitHub workflow (#8)
* Refactor release creation conditions in GitHub workflow

Updated the conditions for invoking the 'Create Release' step in the Github workflow. Now, it triggers only when the 'Get Commits' step is successful. Also fixed a misnamed step reference ('comparison-link' to 'compare-link').

* Update CHANGELOG.md

* Remove redundant date entry from CHANGELOG.md

The commit removes an extra date entry ("2024-02-24") that was added in the changelog file. This cleanup improves the accuracy of the project history tracking by accurately maintaining the changes made on specific dates.

* Create a new .coderabbit.yaml configuration file.

This configuration file sets-up the preferred language, enables early access, and configures the review process and chat settings. It also stipulates conditions for the auto-review feature, including which branches to consider and which titles to ignore.

---------

Co-authored-by: codiumai-pr-agent-pro[bot] <151058649+codiumai-pr-agent-pro[bot]@users.noreply.github.com>