(Changed) Update Configuration Settings in .pr_agent.toml (#11)
* Update incremental review thresholds and changelog push

The incremental review thresholds have been updated in the PR agent configuration, with minimum commits set to 2 and minimum minutes to 10. The option to automatically push changelog changes has been disabled.

* Remove CHANGELOG.md

This commit removes the CHANGELOG.md file from the project. Having the change log in the codebase was found to be redundant because detailed record of changes is already kept in the commit history, improving the clean-up and organization of the project.

---------

Co-authored-by: codiumai-pr-agent-pro[bot] <151058649+codiumai-pr-agent-pro[bot]@users.noreply.github.com>