(Changed) Refactoring Tests for Better Readability and Maintainability (#15)
* Replace hardcoded date and time formats with constants in IsValidDateTest

In the IsValidDateTest unit test, we updated the date and time format strings to use class constants rather than hard-coded values. This makes the code more maintainable and avoids the potential coding errors due to the repetition.

* Refactor variable name in isValidMinute test

Variable name '$Minute' has been changed to lowercase '$minute' for adherence to coding conventions. This change makes sure that all variable names in the test of the 'isValidMinute' method follow consistent naming rules.

* Replace 'hello world' literals with constant in tests

This commit replaces the hard-coded 'hello world' literals in different methods of the SearchWordsTest class with a newly created constant, HELLO_WORLD_LOWERCASE. This modification improves readability and maintainability.

* Refactor StrReplaceTest for better readability

The changes in this commit primarily include the usage of a constant to replace a repeated string in StrReplaceTest. This helps to improve the readability and maintainability of the test code, specifically in the testStrReplaceWithNotFoundSearch and testStrReplace methods.

* Introduce constant for default trim characters

The code has been refactored to use a constant, DEFAULT_TRIM_CHARACTERS, instead of repeating the same set of characters in multiple places. Utilizing a constant enhances readability and maintainability of the code, especially if these trim characters need to be changed in the future.