(Added) Introduce .gitattributes for efficient package management (#34)
* chore: Add .gitattributes to exclude files from archives

- Exclude test directories, meta files, and configuration files
- Reduce package size for end-users and simplify deployment
- Align with best practices for managing Composer packages
- Improve efficiency of installations using `--prefer-dist`

Introduced a .gitattributes file to specify files and directories that should be excluded from distribution archives, including:

- Tests, to lighten the package for production use
- Development-specific files (.gitignore, .github, etc.)
- Configuration files for development tools (PHPUnit, PHPStan, etc.)

This change aims to reduce the package size for end-users, streamline the deployment process, and enhance the efficiency of installations using the `--prefer-dist` option in Composer. It aligns with best practices for managing Composer packages.

* chore: Exclude composer.lock from exported archives

Add '/composer.lock' to .gitattributes export-ignore. This ensures that consumers of the library use their own 'composer.lock' file and have the latest versions of dependencies based on their project's requirements.

* chore: Normalize line endings using .gitattributes

Configure Git to automatically convert line endings to LF on checkin and to CRLF on checkout for Windows. This ensures consistent line endings across different operating systems.

Add the following rule to .gitattributes:

*                   text=auto